+++ tt_metal/fabric/CMakeLists.txt
@@ -44,6 +44,7 @@ target_sources(
         fabric_tensix_builder.cpp
         compressed_routing_table.cpp
         serialization/intermesh_link_table.cpp
+        serialization/physical_desc.cpp
         erisc_datamover_builder_helper.cpp
         ccl/ccl_common.cpp
         physical_system_descriptor.cpp
@@ -53,7 +54,7 @@ target_sources(
 include(flatbuffers)
 set(FLATBUFFER_SCHEMAS
     ${CMAKE_CURRENT_SOURCE_DIR}/serialization/intermesh_link_table.fbs
-    ${CMAKE_CURRENT_SOURCE_DIR}/serialization/exit_node_connection.fbs
+    ${CMAKE_CURRENT_SOURCE_DIR}/serialization/physical_desc.fbs
 )
 
 # Collect all generated headers
