+++ tt_metal/fabric/serialization/physical_desc.fbs
@@ -1,0 +1,78 @@
+namespace tt.tt_metal.flatbuffer;
+
+table EthConnection {
+    src_chan: uint64;
+    dst_chan: uint64;
+    is_local: bool;
+}
+table AsicConnectionEdge {
+    dst_asic_id: uint64;
+    eth_connections: [EthConnection];
+}
+
+table AsicConnnections {
+    asic_connections: [AsicConnectionEdge];
+}
+
+table AsicGraph {
+    asic_id: uint64;
+    topology: AsicConnnections;
+}
+
+table HostAsicConnectivity {
+    host_name: string;
+    asic_topologies: [AsicGraph];
+}
+
+table ExitNodeConnection {
+    src_exit_node: uint64;
+    dst_exit_node: uint64;
+    eth_conn: EthConnection;
+}
+
+table HostConnectionEdge {
+    dst_host_name: string;
+    exit_node_connections: [ExitNodeConnection];
+}
+
+table HostConnections {
+    src_host_name: string;
+    host_connections: [HostConnectionEdge];
+}
+
+table PhysicalConnectivityGraph {
+    asic_connectivity_graph: [HostAsicConnectivity];
+    host_connectivity_graph: [HostConnections];
+}
+
+table AsicDescriptor {
+    tray_id: uint32;
+    n_id: uint32;
+    board_type: uint32;
+    unique_id: uint64;
+    host_name: string;
+}
+
+table AsicDescriptorMap {
+    asic_id: uint64;
+    descriptor: AsicDescriptor;
+}
+
+table HostToMoboMap {
+    host_name: string;
+    mobo_name: string;
+}
+
+table ExitNodeConnectionTable {
+    host_name: string;
+    exit_connections: [ExitNodeConnection];
+}
+
+table PhysicalSystemDescriptor {
+    system_graph: PhysicalConnectivityGraph;
+    asic_descriptors: [AsicDescriptorMap];
+    host_to_mobo_name: [HostToMoboMap];
+    exit_node_connection_table: [ExitNodeConnectionTable];
+}
+
+root_type PhysicalSystemDescriptor;
