+++ tt_metal/fabric/serialization/physical_desc.hpp
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "tt_metal/fabric/physical_system_descriptor.hpp"
+
+namespace tt::tt_metal {
+
+std::vector<uint8_t> serialize_physical_descriptor_to_bytes(
+    const tt_metal::PhysicalSystemDescriptor& physical_descriptor);
+tt_metal::PhysicalSystemDescriptor deserialize_physical_descriptor_from_bytes(const std::vector<uint8_t>& data);
+
+}  // namespace tt::tt_metal
